backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
backoff==2.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
datasets==3.2.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hayhooks==0.0.19
haystack-ai==2.8.0
haystack-experimental==0.4.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
lazy_imports==0.4.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.1
openai==1.59.5
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
posthog==3.7.5
propcache==0.2.1
protobuf==5.29.3
pyarrow==18.1.0
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.12.2
qdrant-haystack==7.0.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
xxhash==3.5.0
yarl==1.18.3
